Reduce size and number of layers
The best practices for Dockerfiles state that the number of layers
should be minimized as long as it doesn't negatively influence
maintainability.
I aggregated the RUN statements a little and grouped them into
mostly coherent blocks and tried to minimize the number of
temporary files that make it into a layer of the incremental file system
(as those will always be present in the layers and only be marked as
deleted).
I also only installed the SDK versions for which an emulator image
is installed in order to reduce file size.
The original image had about 9.5 GB and if you use devicemapper2 as
storage engine there is a 10GB size limit which occasionally led to
"No disk space left" errors.